[refact] 토큰 재발급 리팩토링
