+++ src/main/java/kusitms/backend/domain/token/dto/response/TokenResponse.java
@@ -1,11 +1,13 @@
-package kusitms.backend.domain.refreshtoken.dto.response;
+package kusitms.backend.domain.token.dto.response;
 
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
+@Builder
 public class TokenResponse {
     private String accessToken;
     private String refreshToken;
