+++ src/main/java/kusitms/backend/domain/token/repository/RefreshTokenRepository.java
@@ -1,8 +1,9 @@
-package kusitms.backend.domain.refreshtoken.repository;
+package kusitms.backend.domain.token.repository;
 
-import kusitms.backend.domain.refreshtoken.entity.RefreshToken;
+import kusitms.backend.domain.token.entity.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
     RefreshToken findByUserId(Long userId);
+    boolean existsByToken(String refreshToken);
 }
